CMS PM2 separate file
