+++ cms/src/pm2.js
@@ -1,0 +1,14 @@
+import pm2Config from './../pm2.config.js';
+
+const pm2Env = process.env.ENV;
+if (!pm2Env) {
+  throw new Error('No ENV value provided!');
+}
+const pm2ConfigGeneric =
+  (pm2Config && pm2Config.apps && pm2Config.apps[0] && pm2Config.apps[0].env) || {};
+const pm2ConfigForEnv =
+  (pm2Config && pm2Config.apps && pm2Config.apps[0] && pm2Config.apps[0][`env_${pm2Env}`]) || {};
+
+const pm2 = { ...pm2ConfigGeneric, ...pm2ConfigForEnv };
+
+export default pm2;
