+++ cms/src/services/search-client/indexModel.js
@@ -1,6 +1,6 @@
 // @ts-check
 import getLogger from '../../logger.js';
-import { pm2 } from '../../index.js';
+import pm2 from '../../pm2.js';
 
 import getClient from './client.js';
